Changes and a test for directly-recursive data structures.  I.E. structures of the form:
typedef struct node {
    int node_num;
    struct node *link1;
    struct node *link2;
} *node_ptr;

FMField node_field_list[] =
{
    {"node_num", "integer", sizeof(int), FMOffset(node_ptr, node_num)},
    {"link1", "*node", sizeof(struct node), FMOffset(node_ptr, link1)},
    {"link2", "*node", sizeof(struct node), FMOffset(node_ptr, link2)},
    {(char *) 0, (char *) 0, 0, 0}
};

Context test now transmits three structures, a list with a first/last link,
a graph with a single loop, and a binary tree.   On the transmit side we
calculate a "signature" of the data structure, a function of the node_num
and the depth and position at which it is encountered.  This signature
replaces the node_num of the first node.  On the receive side, we replace
that node_num with zero and calculate the sig again.  If they don't match,
we have a problem.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@3394 8ba6a8b9-845b-0410-94f0-974be2f5df47