We've always tested to see if the output files produced by context_test and write_file_output were different than those produced by prior runs for the particular architecture.  This has gotten to be a pain as we've added different flavors of linux, etc. that all have the exact same output files, but we have to add file for the architecture, etc.
This switches to a different mechanism.  Rather than comparing the new
output file (with a specific name based on architecture) with an old one (in
ffs/output_files and with a specific name based on architecture), we instead
calculate file size and checksum of the current file and compare it to all
the file sizes and checksums of the files in the output directory.  If the
current file has the same size and checksum as something in the output
directory, everything is good and we say we passed.  If the current one is
unique, either the output format has changed (!) or this is a new unique
architecture and we should have an output file from it in the output_files
dir.  Fail until one is added there.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@13760 8ba6a8b9-845b-0410-94f0-974be2f5df47