Modify code generation so that we reasonably generate code for CoD constructs of the form "const int var = init;".  In particular, we treat it as a compile-time constant for which we generate code only at the time of usage.
Drop some header comments in the tops of some tests...

Revive t7.  This is not a proper regression test, but is only a bit of code
that has been used now and then to manually check the operation of things
that aren't easy to observe without looking at generated code.  The DCG for
"const int var" stuff is exercised there.


git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@17465 8ba6a8b9-845b-0410-94f0-974be2f5df47